move some operations from enrollee to kit classes
