+++ api-admin/src/main/java/bio/terra/pearl/api/admin/controller/enrollee/EnrolleeController.java
@@ -5,9 +5,7 @@ import bio.terra.pearl.api.admin.service.AuthUtilService;
 import bio.terra.pearl.api.admin.service.enrollee.EnrolleeExtService;
 import bio.terra.pearl.core.model.EnvironmentName;
 import bio.terra.pearl.core.model.admin.AdminUser;
-import bio.terra.pearl.core.model.kit.KitRequest;
 import bio.terra.pearl.core.model.participant.Enrollee;
-import bio.terra.pearl.core.service.kit.PepperException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.UUID;
 import javax.servlet.http.HttpServletRequest;
@@ -59,33 +57,6 @@ public class EnrolleeController implements EnrolleeApi {
     } catch (JsonProcessingException e) {
       return ResponseEntity.internalServerError().body(e.getMessage());
     }
-  }
-
-  @Override
-  public ResponseEntity<Object> requestKit(
-      String portalShortcode,
-      String studyShortcode,
-      String envName,
-      String enrolleeShortcode,
-      String kitType) {
-    AdminUser adminUser = authUtilService.requireAdminUser(request);
-    try {
-      KitRequest sampleKit = enrolleeExtService.requestKit(adminUser, enrolleeShortcode, kitType);
-      return ResponseEntity.ok(sampleKit);
-    } catch (PepperException e) {
-      log.error("Error requesting sample kit from Pepper", e);
-      // In the case of a PepperException, we can do better than this because we'll likely know what
-      // Pepper was unhappy about, such as an address failing to validate.
-      return ResponseEntity.internalServerError().body(e.getMessage());
-    }
-  }
-
-  @Override
-  public ResponseEntity<Object> getKitRequests(
-      String portalShortcode, String studyShortcode, String envName, String enrolleeShortcode) {
-    AdminUser adminUser = authUtilService.requireAdminUser(request);
-    var kitRequests = enrolleeExtService.getKitRequests(adminUser, enrolleeShortcode);
-    return ResponseEntity.ok(kitRequests);
   }
 
   @Override
