+++ api-admin/src/main/java/bio/terra/pearl/api/admin/controller/kit/KitController.java
@@ -5,6 +5,8 @@ import bio.terra.pearl.api.admin.service.AuthUtilService;
 import bio.terra.pearl.api.admin.service.kit.KitExtService;
 import bio.terra.pearl.core.model.EnvironmentName;
 import bio.terra.pearl.core.model.admin.AdminUser;
+import bio.terra.pearl.core.model.kit.KitRequest;
+import bio.terra.pearl.core.service.kit.PepperException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Arrays;
 import javax.servlet.http.HttpServletRequest;
@@ -40,6 +42,33 @@ public class KitController implements KitApi {
             adminUser, portalShortcode, studyShortcode, environmentName);
 
     return ResponseEntity.ok(kits);
+  }
+
+  @Override
+  public ResponseEntity<Object> requestKit(
+      String portalShortcode,
+      String studyShortcode,
+      String envName,
+      String enrolleeShortcode,
+      String kitType) {
+    AdminUser adminUser = authUtilService.requireAdminUser(request);
+    try {
+      KitRequest sampleKit = kitExtService.requestKit(adminUser, enrolleeShortcode, kitType);
+      return ResponseEntity.ok(sampleKit);
+    } catch (PepperException e) {
+      log.error("Error requesting sample kit from Pepper", e);
+      // In the case of a PepperException, we can do better than this because we'll likely know what
+      // Pepper was unhappy about, such as an address failing to validate.
+      return ResponseEntity.internalServerError().body(e.getMessage());
+    }
+  }
+
+  @Override
+  public ResponseEntity<Object> getKitRequests(
+      String portalShortcode, String studyShortcode, String envName, String enrolleeShortcode) {
+    AdminUser adminUser = authUtilService.requireAdminUser(request);
+    var kitRequests = kitExtService.getKitRequests(adminUser, enrolleeShortcode);
+    return ResponseEntity.ok(kitRequests);
   }
 
   @Override
