+++ api-admin/src/main/java/bio/terra/pearl/api/admin/service/enrollee/EnrolleeExtService.java
@@ -3,13 +3,11 @@ package bio.terra.pearl.api.admin.service.enrollee;
 import bio.terra.pearl.api.admin.service.AuthUtilService;
 import bio.terra.pearl.core.model.EnvironmentName;
 import bio.terra.pearl.core.model.admin.AdminUser;
-import bio.terra.pearl.core.model.kit.KitRequest;
 import bio.terra.pearl.core.model.participant.Enrollee;
 import bio.terra.pearl.core.model.participant.EnrolleeSearchResult;
 import bio.terra.pearl.core.model.participant.WithdrawnEnrollee;
 import bio.terra.pearl.core.model.workflow.DataChangeRecord;
 import bio.terra.pearl.core.service.exception.PermissionDeniedException;
-import bio.terra.pearl.core.service.kit.KitRequestService;
 import bio.terra.pearl.core.service.participant.EnrolleeService;
 import bio.terra.pearl.core.service.participant.WithdrawnEnrolleeService;
 import bio.terra.pearl.core.service.participant.search.EnrolleeSearchService;
@@ -18,7 +16,6 @@ import bio.terra.pearl.core.service.portal.PortalService;
 import bio.terra.pearl.core.service.study.PortalStudyService;
 import bio.terra.pearl.core.service.workflow.DataChangeRecordService;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import java.util.Collection;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
@@ -30,7 +27,6 @@ public class EnrolleeExtService {
   private PortalService portalService;
   private PortalStudyService portalStudyService;
   private DataChangeRecordService dataChangeRecordService;
-  private KitRequestService kitRequestService;
 
   private EnrolleeSearchService enrolleeSearchService;
 
@@ -41,7 +37,6 @@ public class EnrolleeExtService {
       PortalService portalService,
       PortalStudyService portalStudyService,
       DataChangeRecordService dataChangeRecordService,
-      KitRequestService kitRequestService,
       EnrolleeSearchService enrolleeSearchService) {
     this.authUtilService = authUtilService;
     this.enrolleeService = enrolleeService;
@@ -49,7 +44,6 @@ public class EnrolleeExtService {
     this.portalService = portalService;
     this.portalStudyService = portalStudyService;
     this.dataChangeRecordService = dataChangeRecordService;
-    this.kitRequestService = kitRequestService;
     this.enrolleeSearchService = enrolleeSearchService;
   }
 
@@ -89,15 +83,5 @@ public class EnrolleeExtService {
     }
     Enrollee enrollee = authUtilService.authAdminUserToEnrollee(user, enroleeShortcode);
     return withdrawnEnrolleeService.withdrawEnrollee(enrollee);
-  }
-
-  public KitRequest requestKit(AdminUser adminUser, String enrolleeShortcode, String kitTypeName) {
-    Enrollee enrollee = authUtilService.authAdminUserToEnrollee(adminUser, enrolleeShortcode);
-    return kitRequestService.requestKit(adminUser, enrollee, kitTypeName);
-  }
-
-  public Collection<KitRequest> getKitRequests(AdminUser adminUser, String enrolleeShortcode) {
-    Enrollee enrollee = authUtilService.authAdminUserToEnrollee(adminUser, enrolleeShortcode);
-    return kitRequestService.getKitRequests(enrollee);
   }
 }
