+++ api-admin/src/main/java/bio/terra/pearl/api/admin/service/kit/KitExtService.java
@@ -1,10 +1,10 @@
 package bio.terra.pearl.api.admin.service.kit;
 
 import bio.terra.pearl.api.admin.service.AuthUtilService;
-import bio.terra.pearl.api.admin.service.enrollee.EnrolleeExtService;
 import bio.terra.pearl.core.model.EnvironmentName;
 import bio.terra.pearl.core.model.admin.AdminUser;
 import bio.terra.pearl.core.model.kit.KitRequest;
+import bio.terra.pearl.core.model.participant.Enrollee;
 import bio.terra.pearl.core.model.study.StudyEnvironment;
 import bio.terra.pearl.core.service.kit.KitRequestService;
 import bio.terra.pearl.core.service.study.StudyEnvironmentService;
@@ -15,17 +15,14 @@ import org.springframework.stereotype.Service;
 @Service
 public class KitExtService {
   private final AuthUtilService authUtilService;
-  private final EnrolleeExtService enrolleeExtService;
   private final KitRequestService kitRequestService;
   private final StudyEnvironmentService studyEnvironmentService;
 
   public KitExtService(
       AuthUtilService authUtilService,
-      EnrolleeExtService enrolleeExtService,
       KitRequestService kitRequestService,
       StudyEnvironmentService studyEnvironmentService) {
     this.authUtilService = authUtilService;
-    this.enrolleeExtService = enrolleeExtService;
     this.kitRequestService = kitRequestService;
     this.studyEnvironmentService = studyEnvironmentService;
   }
@@ -40,9 +37,7 @@ public class KitExtService {
     authUtilService.authUserToStudy(adminUser, portalShortcode, studyShortcode);
 
     return enrolleeShortcodes.stream()
-        .map(
-            enrolleeShortcode ->
-                enrolleeExtService.requestKit(adminUser, enrolleeShortcode, kitType))
+        .map(enrolleeShortcode -> requestKit(adminUser, enrolleeShortcode, kitType))
         .toList();
   }
 
@@ -57,5 +52,15 @@ public class KitExtService {
         studyEnvironmentService.findByStudy(studyShortcode, environmentName).get();
 
     return kitRequestService.getSampleKitsByStudyEnvironment(studyEnvironment);
+  }
+
+  public KitRequest requestKit(AdminUser adminUser, String enrolleeShortcode, String kitTypeName) {
+    Enrollee enrollee = authUtilService.authAdminUserToEnrollee(adminUser, enrolleeShortcode);
+    return kitRequestService.requestKit(adminUser, enrollee, kitTypeName);
+  }
+
+  public Collection<KitRequest> getKitRequests(AdminUser adminUser, String enrolleeShortcode) {
+    Enrollee enrollee = authUtilService.authAdminUserToEnrollee(adminUser, enrolleeShortcode);
+    return kitRequestService.getKitRequests(enrollee);
   }
 }
